Update to io.ebean.
org.ajave no longer supports modern Java versions, but the failures only
happen at runtime. io.ebean is the successor to this project, and is
maybe 90% drop in replacement for the older versions, with the one major
exception that the database creation is no longer supported, hence the
deltion of EbeanHelper.java. Other than that, the remaining top level
methods have all been updated and are drop in replacements.